+++ examples/shared_runtime/queries/module_1.sql
@@ -1,0 +1,3 @@
+--! insert_book
+INSERT INTO books (title)
+  VALUES (:title);